feat: JoinRevenueModal add button
